# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/staff": {
      "get": {
        "operationId": "StaffController_list",
        "parameters": [
          {
            "name": "storeId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StaffMemberDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff"
        ]
      },
      "post": {
        "operationId": "StaffController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStaffDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffMemberDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff"
        ]
      }
    },
    "/v1/staff/{id}": {
      "patch": {
        "operationId": "StaffController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStaffDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffMemberDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff"
        ]
      }
    },
...
  },
  "info": {
    "title": "LORA API",
    "description": "White-label booking platform API",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "StaffMemberDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "storeId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "jobTitle": {
            "type": "object",
            "nullable": true
          },
          "photoUrl": {
            "type": "object",
            "nullable": true
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "storeId",
          "name",
          "jobTitle",
          "photoUrl",
          "active"
        ]
      },
      "CreateStaffDto": {
        "type": "object",
        "properties": {
          "storeId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Aisha Rahman"
          },
          "jobTitle": {
            "type": "string",
            "example": "Senior Therapist"
          },
          "photoUrl": {
            "type": "string"
          },
          "active": {
            "type": "boolean",
            "description": "Defaults to active."
          }
        },
        "required": [
          "storeId",
          "name"
        ]
      },
      "UpdateStaffDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "jobTitle": {
            "type": "string"
          },
          "photoUrl": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}